pyproject.toml
[project]
name = "accessdane-audit"
version = "0.1.0"
description = "AccessDane McFarland audit pipeline"
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.13",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "sqlalchemy>=2.0",
  "typer>=0.12",
  "python-dotenv>=1.0",
  "psycopg>=3.1",
]

[project.scripts]
accessdane = "accessdane_audit.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
